backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
faster-whisper==1.0.3
httpx==0.27.0
python-multipart==0.0.9
yake==0.4.8
deep-translator==1.11.4
Pillow==10.3.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.131.0
itsdangerous==2.2.0
